[WIP]scripts: Add json_sbom to generate SBOM during build
This commit adds the json_sbom script, which enables the generation
of a Software Bill of Materials (SBOM) during the OpenWrt build process.
The SBOM provides a comprehensive list of all software components and
their dependencies used in the build, allowing for improved transparency
and security in the software supply chain.

By including the json_sbom script, OpenWrt users and developers can
easily generate an SBOM file that captures the relevant information
about the software stack.
This SBOM can then be used for various purposes, such as vulnerability
management, license compliance, and ensuring the traceability of
software components

Signed-off-by: Oliver Sedlbauer <osedlbauer@tdt.de>